Feed enhancements (#28)
* repositioned save and report icons on QuestionPage

* fix for long question titles pushing save/report icons to next line.

* implemented feed memory (tracks the last post you viewed, only shows those you haven't seen + new posts). Can view older posts when reached end of new posts and click "Refresh Feed"

* implemented going back to original feed and in same location after viewing community, simultaneously accounting for feed memory. now I'm going to do the same for when user clicks on view post.

* implemented view post from feed and go back to same feed and position.

* updated report limit, added floating banner message for interest/not interested/report to undo the action, changed color of checkmark for refresh feed.

* implemented fading undo message for report from answer page

* modified deletion functionality for certain types:

deleting a user now deletes it from the communities it was a part of

deleting a question now deletes its answers and comments (but not replies to the comments, which would require recursion)

upon question deletion, we clean out tags related to this question but not used in other questions.

* updated deletion functionality to make sure to remove a question from a user's savedQuestions if it's been deleted. Also, if a tag has been deleted as a result of deleting a question, all users' interests are updated to not include the interest corresponding to that tag anymore.

* fixed dependency issues, modified user deletion to make sure user is removed from its communities as well (updating counts on community cards). Likewise removed question from its communities upon deletion.

* fixed dependency issue

* fixed failing test for deleting user.

* commented out code related to sharing popup

+++ server/services/question.service.ts
@@ -18,7 +18,7 @@ import QuestionModel from '../models/questions.model';
 import TagModel from '../models/tags.model';
 import CommentModel from '../models/comments.model';
 import { parseKeyword, parseTags } from '../utils/parse.util';
-import { checkTagInQuestion } from './tag.service';
+import { checkTagInQuestion, deleteTagsByIds } from './tag.service';
 import {
   sortQuestionsByActive,
   sortQuestionsByMostViews,
@@ -26,6 +26,10 @@ import {
   sortQuestionsByUnanswered,
 } from '../utils/sort.util';
 import CommunityModel from '../models/communities.model';
+import { deleteAnswerById } from './answer.service';
+import { deleteCommentById } from './comment.service';
+import { deleteInterestByTagId } from './interest.service';
+import UserModel from '../models/users.model';
 
 /**
  * Checks if keywords exist in a question's title or text.
@@ -175,8 +179,78 @@ export const saveQuestion = async (question: Question): Promise<QuestionResponse
  */
 export const deleteQuestionById = async (qid: string): Promise<QuestionResponse> => {
   try {
+    // Populate to get the comments
+    const question: PopulatedDatabaseQuestion | null = await QuestionModel.findById(qid).populate<{
+      tags: DatabaseTag[];
+      answers: PopulatedDatabaseAnswer[];
+      comments: DatabaseComment[];
+    }>([
+      { path: 'tags', model: TagModel },
+      { path: 'answers', model: AnswerModel, populate: { path: 'comments', model: CommentModel } },
+      { path: 'comments', model: CommentModel },
+    ]);
+
+    if (!question) {
+      return { error: 'Question not found' };
+    }
+
     const result: DatabaseQuestion | null = await QuestionModel.findByIdAndDelete(qid);
-    return result || { error: 'Question not found' };
+
+    if (!result) {
+      return { error: 'Question not found' };
+    }
+
+    // Delete comments to answers
+    const deleteAnswerCommentPromises = question.answers.map(answer =>
+      answer.comments.map(comment => deleteCommentById(comment._id.toString())),
+    );
+    await Promise.all(deleteAnswerCommentPromises);
+
+    // Delete all answers related to the question, using deleteAnswerById
+    const deleteAnswerPromises = result.answers.map(answer =>
+      deleteAnswerById(answer._id.toString()),
+    );
+    await Promise.all(deleteAnswerPromises);
+
+    // Delete all comments related to the question, using deleteCommentById
+    const deleteCommentPromises = result.comments.map(comment =>
+      deleteCommentById(comment._id.toString()),
+    );
+    await Promise.all(deleteCommentPromises);
+
+    // For each tag, check if any other question is using it. If not, append to the list of tags to be deleted
+    const tagsToDelete: ObjectId[] = [];
+    const tagExistenceChecks = result.tags.map(tag => QuestionModel.exists({ tags: tag }));
+    const tagExistenceResults = await Promise.all(tagExistenceChecks);
+
+    result.tags.forEach((tag, index) => {
+      if (!tagExistenceResults[index]) {
+        tagsToDelete.push(tag);
+      }
+    });
+
+    if (tagsToDelete.length === 0) {
+      return result;
+    }
+
+    // Delete all interests related to the tags
+    const deleteInterestPromises = tagsToDelete.map(tagId => deleteInterestByTagId(tagId));
+    await Promise.all(deleteInterestPromises);
+
+    // Delete the tags that are not used by any other question
+    await deleteTagsByIds(tagsToDelete);
+
+    // Remove from savedQuestions
+    await UserModel.updateMany({ savedQuestions: qid }, { $pull: { savedQuestions: qid } });
+
+    // Get the community of the question and remove it from that community's questions
+    const communities = await CommunityModel.find({ questions: qid });
+    const communityPromises = communities.map(community =>
+      CommunityModel.findByIdAndUpdate(community._id, { $pull: { questions: qid } }, { new: true }),
+    );
+    await Promise.all(communityPromises);
+
+    return result;
   } catch (error) {
     return { error: 'Error when deleting a question' };
   }
